wip: get the new component to show in the dr-ui docs
